Edmx update(Identity Management) - In Progress

+++ LYSAdmin.Web/LYSAdmin.Data/DBEntity/UserDetail.cs
@@ -15,7 +15,7 @@ namespace LYSAdmin.Data.DBEntity
     public partial class UserDetail
     {
         public int UserDetailsID { get; set; }
-        public int UserID { get; set; }
+        public long UserID { get; set; }
         public string PresentAddress { get; set; }
         public string PermanentAddress { get; set; }
         public int GovtIDType { get; set; }
